Refactor: 애니메이션 수정, 버그 수정
애니메이션을 자연스럽게, 로직을 단순하게 수정했습니다

+++ Assets/02_Scripts/Enemy/Data/EnemySO.cs
@@ -11,6 +11,6 @@ public class EnemySO : ScriptableObject
     public float _speed;
     public float _moveDelay;
     public float _attackRange;
-    public float _attackDamage;
+    public int _attackDamage;
     public float _attackDelay;
 }
